Refactor somatic pipeline
