+++ janis_pipelines/__init__.py
@@ -10,12 +10,11 @@ from janis_pipelines.wgs_germline_gatk.wgsgermlinegatk import WGSGermlineGATK
 from janis_pipelines.wgs_germline_gatk.wgsgermlinegatk_variantsonly import (
     WGSGermlineGATKVariantsOnly,
 )
-
-from janis_pipelines.wgs_somatic.wgssomatic import (
-    WGSSomaticMultiCallers,
+from janis_pipelines.wgs_somatic.wgssomatic import WGSGermlineMultiCallers
+from janis_pipelines.wgs_somatic.wgssomatic_variantsonly import (
     WGSSomaticMultiCallersVariantsOnly,
 )
-from janis_pipelines.wgs_somatic_gatk.wgssomaticgatk import (
-    WGSSomaticGATK,
+from janis_pipelines.wgs_somatic_gatk.wgssomaticgatk import WGSSomaticGATK
+from janis_pipelines.wgs_somatic_gatk.wgssomaticgatk_variantsonly import (
     WGSSomaticGATKVariantsOnly,
 )
