+++ janis_pipelines/wgs_somatic_gatk/wgssomaticgatk_variantsonly.py
@@ -19,61 +19,13 @@ from janis_bioinformatics.data_types import (
     CompressedVcf,
     Vcf,
 )
-from janis_bioinformatics.tools.bcftools import BcfToolsSort_1_9, BcfToolsConcat_1_9
-from janis_bioinformatics.tools.bioinformaticstoolbase import BioinformaticsWorkflow
-from janis_bioinformatics.tools.common import GATKBaseRecalBQSRWorkflow_4_1_3
-from janis_bioinformatics.tools.pmac import (
-    AddBamStatsSomatic_0_1_0,
-    GenerateIntervalsByChromosome,
+from janis_pipelines.wgs_somatic_gatk.wgssomaticgatk import WGSSomaticGATK
+from janis_pipelines.wgs_somatic.wgssomatic_variantsonly import (
+    WGSSomaticMultiCallersVariantsOnly,
 )
-from janis_bioinformatics.tools.variantcallers import GatkSomaticVariantCaller_4_1_3
-
-from janis_pipelines.reference import WGS_INPUTS
-
-INPUT_DOCS = {
-    **WGS_INPUTS,
-    "normal_inputs": {
-        "doc": "An array of NORMAL FastqGz pairs. These are aligned separately and merged "
-        "to create higher depth coverages from multiple sets of reads",
-        "quality": InputQualityType.user,
-        "example": [
-            ["normal_R1.fastq.gz", "normal_R2.fastq.gz"],
-            ["normal_R1-TOPUP.fastq.gz", "normal_R2-TOPUP.fastq.gz"],
-        ],
-    },
-    "tumor_inputs": {
-        "doc": "An array of TUMOR FastqGz pairs. These are aligned separately and merged "
-        "to create higher depth coverages from multiple sets of reads",
-        "quality": InputQualityType.user,
-        "example": [
-            ["tumor_R1.fastq.gz", "tumor_R2.fastq.gz"],
-            ["tumor_R1-TOPUP.fastq.gz", "tumor_R2-TOPUP.fastq.gz"],
-        ],
-    },
-    "normal_name": {
-        "doc": "Sample name for the NORMAL sample from which to generate the readGroupHeaderLine for BwaMem",
-        "quality": InputQualityType.user,
-        "example": "NA12878_normal",
-    },
-    "tumor_name": {
-        "doc": "Sample name for the TUMOR sample from which to generate the readGroupHeaderLine for BwaMem",
-        "quality": InputQualityType.user,
-        "example": "NA12878_tumor",
-    },
-    "normal_bam": {
-        "doc": "Indexed NORMAL bam to call somatic variants against",
-        "quality": InputQualityType.user,
-        "example": "NA12878-normal.bam",
-    },
-    "tumor_bam": {
-        "doc": "Indexed TUMOR bam to call somatic variants against",
-        "quality": InputQualityType.user,
-        "example": "NA12878-normal.bam",
-    },
-}
 
 
-class WGSSomaticGATKVariantsOnly(BioinformaticsWorkflow):
+class WGSSomaticGATKVariantsOnly(WGSSomaticGATK, WGSSomaticMultiCallersVariantsOnly):
     def id(self):
         return "WGSSomaticGATKVariantsOnly"
 
@@ -84,166 +36,44 @@ class WGSSomaticGATKVariantsOnly(BioinformaticsWorkflow):
         return "1.4.0"
 
     def constructor(self):
-
         self.add_inputs()
+        self.add_localise_reference()
+        self.add_localise_bam()
+        self.add_bam_qc(
+            normal_bam_source=self.localise_normal_bam.out,
+            tumor_bam_source=self.localise_tumor_bam.out,
+        )
         self.add_gatk_variantcaller(
-            normal_bam_source=self.normal_bam, tumor_bam_source=self.tumor_bam
+            normal_bam_source=self.localise_normal_bam.out,
+            tumor_bam_source=self.localise_tumor_bam.out,
         )
         self.add_addbamstats(
-            normal_bam_source=self.normal_bam, tumor_bam_source=self.tumor_bam
+            normal_bam_source=self.localise_normal_bam.out,
+            tumor_bam_source=self.localise_tumor_bam.out,
         )
 
     def add_inputs(self):
-
         # INPUTS
-        self.input("normal_bam", BamBai, doc=INPUT_DOCS["normal_bam"])
-        self.input("tumor_bam", BamBai, doc=INPUT_DOCS["tumor_bam"])
-        self.input("normal_name", String(), doc=INPUT_DOCS["normal_name"])
-        self.input("tumor_name", String(), doc=INPUT_DOCS["tumor_name"])
+        self.input("normal_bam", BamBai, doc="normal_bam")
+        self.input("tumor_bam", BamBai, doc="tumor_bam")
+        self.input("normal_name", String, doc="normal_name")
+        self.input("tumor_name", String, doc="tumor_name")
+        # self.input("normal_bam", BamBai, doc=INPUT_DOCS["normal_bam"])
+        # self.input("tumor_bam", BamBai, doc=INPUT_DOCS["tumor_bam"])
+        # self.input("normal_name", String, doc=INPUT_DOCS["normal_name"])
+        # self.input("tumor_name", String, doc=INPUT_DOCS["tumor_name"])
 
-        self.add_inputs_for_configuration()
         self.add_inputs_for_reference()
+        self.add_inputs_for_adapter_trimming()
         self.add_inputs_for_intervals()
-
-    def add_inputs_for_configuration(self):
-
-        self.input("gnomad", VcfTabix(), doc=INPUT_DOCS["gnomad"])
-        self.input(
-            "panel_of_normals",
-            VcfTabix(optional=True),
-            doc=INPUT_DOCS["panel_of_normals"],
-        )
-
-    def add_inputs_for_intervals(self):
-        self.input("gatk_intervals", Array(Bed), doc=INPUT_DOCS["gatk_intervals"])
-        self.input("gridss_blacklist", Bed, doc=INPUT_DOCS["gridss_blacklist"])
-
-    def add_inputs_for_reference(self):
-
-        self.input("reference", FastaWithDict, doc=INPUT_DOCS["reference"])
-        self.input("snps_dbsnp", VcfTabix, doc=INPUT_DOCS["snps_dbsnp"])
-        self.input("snps_1000gp", VcfTabix, doc=INPUT_DOCS["snps_1000gp"])
-        self.input("known_indels", VcfTabix, doc=INPUT_DOCS["known_indels"])
-        self.input("mills_indels", VcfTabix, doc=INPUT_DOCS["mills_indels"])
-
-    def add_inputs_for_adapter_trimming(self):
-        self.input("adapter_file", File)
-        self.input("contaminant_file", File)
-
-    def add_gatk_variantcaller(self, normal_bam_source, tumor_bam_source):
-        if "generate_gatk_intervals" in self.step_nodes:
-            generated_intervals = self.generate_gatk_intervals.out_regions
-        else:
-            generated_intervals = self.step(
-                "generate_gatk_intervals",
-                GenerateIntervalsByChromosome(reference=self.reference),
-                when=self.gatk_intervals.is_null(),
-            ).out_regions
-
-        intervals = FirstOperator([self.gatk_intervals, generated_intervals])
-
-        recal_ins = {
-            "reference": self.reference,
-            "intervals": intervals,
-            "snps_dbsnp": self.snps_dbsnp,
-            "snps_1000gp": self.snps_1000gp,
-            "known_indels": self.known_indels,
-            "mills_indels": self.mills_indels,
-        }
-        self.step(
-            "bqsr_normal",
-            GATKBaseRecalBQSRWorkflow_4_1_3(bam=normal_bam_source, **recal_ins),
-            scatter="intervals",
-        )
-
-        self.step(
-            "bqsr_tumor",
-            GATKBaseRecalBQSRWorkflow_4_1_3(bam=tumor_bam_source, **recal_ins),
-            scatter="intervals",
-        )
-
-        self.step(
-            "vc_gatk",
-            GatkSomaticVariantCaller_4_1_3(
-                normal_bam=self.bqsr_normal.out,
-                tumor_bam=self.bqsr_tumor.out,
-                normal_name=self.normal_name,
-                intervals=intervals,
-                reference=self.reference,
-                gnomad=self.gnomad,
-                panel_of_normals=self.panel_of_normals,
-            ),
-            scatter=["intervals", "normal_bam", "tumor_bam"],
-        )
-
-        self.step(
-            "vc_gatk_merge",
-            BcfToolsConcat_1_9(vcf=self.vc_gatk.out.as_type(Array(Vcf))),
-        )
-        self.step(
-            "vc_gatk_sort_combined",
-            BcfToolsSort_1_9(vcf=self.vc_gatk_merge.out.as_type(CompressedVcf)),
-        )
-        self.step(
-            "vc_gatk_uncompressvcf",
-            UncompressArchive(file=self.vc_gatk_sort_combined.out),
-        )
-
-        # VCF
-        self.output(
-            "out_variants_gatk",
-            source=self.vc_gatk_sort_combined.out,
-            output_folder="variants",
-            output_name=StringFormatter(
-                "{tumor_name}--{normal_name}_gatk",
-                tumor_name=self.tumor_name,
-                normal_name=self.normal_name,
-            ),
-            doc="Merged variants from the GATK caller",
-        )
-        self.output(
-            "out_variants_gatk_split",
-            source=self.vc_gatk.out,
-            output_folder=[
-                "variants",
-                "GatkByInterval",
-            ],
-            doc="Unmerged variants from the GATK caller (by interval)",
-        )
-
-    def add_addbamstats(self, normal_bam_source, tumor_bam_source):
-        self.step(
-            "addbamstats",
-            AddBamStatsSomatic_0_1_0(
-                normal_id=self.normal_name,
-                tumor_id=self.tumor_name,
-                normal_bam=normal_bam_source,
-                tumor_bam=tumor_bam_source,
-                reference=self.reference,
-                vcf=self.vc_gatk_uncompressvcf.out.as_type(Vcf),
-            ),
-        )
-
-        self.output(
-            "out_variants_bamstats",
-            source=self.addbamstats.out,
-            output_folder=[
-                "variants",
-            ],
-            output_name=StringFormatter(
-                "{tumor_name}--{normal_name}",
-                tumor_name=self.tumor_name,
-                normal_name=self.normal_name,
-            ),
-            doc="Final vcf from GATK",
-        )
+        self.add_inputs_for_configuration()
 
     def bind_metadata(self):
         meta: WorkflowMetadata = self.metadata
 
         meta.keywords = ["wgs", "cancer", "somatic", "variants", "gatk"]
         meta.dateUpdated = date(2019, 10, 16)
-        meta.dateUpdated = date(2021, 9, 3)
+        meta.dateUpdated = date(2022, 3, 1)
 
         meta.contributors = ["Michael Franklin", "Richard Lupat", "Jiaan Yu"]
         meta.short_documentation = "A somatic tumor-normal variant-calling WGS pipeline using only GATK Mutect2"
@@ -269,25 +99,18 @@ The known sites (snps_dbsnp, snps_1000gp, known_indels, mills_indels) should be 
 
 
 if __name__ == "__main__":
-    import os.path
+    # import os.path
 
-    w = WGSSomaticGATKVariantsOnly()
-    args = {
-        "to_console": False,
-        "to_disk": True,
-        "validate": True,
-        "export_path": os.path.join(
-            os.path.dirname(os.path.realpath(__file__)), "{language}"
-        ),
-    }
-    # w.get_dot_plot(show=True, expand_subworkflows=True)
+    # w = WGSSomaticGATKVariantsOnly()
+    # args = {
+    #     "to_console": False,
+    #     "to_disk": False,
+    #     "validate": True,
+    #     "export_path": os.path.join(
+    #         os.path.dirname(os.path.realpath(__file__)), "{language}"
+    #     ),
+    #     "with_resource_overrides": True,
+    # }
     # w.translate("cwl", **args)
     # w.translate("wdl", **args)
-    #
-    # # from cwltool import main
-    # # import logging
-    #
-    # # op = os.path.dirname(os.path.realpath(__file__)) + "/cwl/WGSGermlineGATK.py"
-    #
-    # # main.run(*["--validate", op], logger_handler=logging.Handler())
-    w.translate("wdl")
+    WGSSomaticGATKVariantsOnly().translate("wdl")
